Camelot Implementation
Not entirely sure this works, haven't run it yet but it seems like this will take a .pdf, break it into Text, Tables, and Images. Text gets added to text document, tables are extracted and given a description and broken into text, and then images are ran through LlaVA. All is added to the text document and then saved as a .txt.